# Back-end/designDocs/API_DOCS.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IRL Paperspace control API",
    "version": "1.0.0"
  },
  "paths": {
    "/wco/startSession": {
      "get": {
        "description": "sets Session id and authenticates",
        "tags": [
          "wco"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "success"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "WCO API Error"
          }
        }
      }
    },
    "/wco/reservations": {
      "get": {
        "description": "lists all reservations from wco api",
        "tags": [
          "wco"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "success"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "WCO API Error"
          }
        }
      }
    },
    "/utils/machines": {
      "post": {
        "description": "returns machines with given state",
        "tags": [
          "paperspaceUtils"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "description": "state",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/State"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "found machines with state"
          },
          "400": {
            "description": "Bad request body"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "Paperspace API Error"
          }
        }
      },
      "get": {
        "description": "gets all machines in team",
        "tags": [
          "paperspaceUtils"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "send machine list"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "Paperspace API Error"
          }
        }
      }
    },
    "/utils/users": {
      "post": {
        "description": "gets users based on email",
        "tags": [
          "paperspaceUtils"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "description": "email",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Email"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successfully got user"
          },
          "400": {
            "description": "Bad request body"
          },
          "404": {
            "description": "user not found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "Paperspace API Error"
          }
        }
      },
      "get": {
        "description": "gets all users in team",
        "tags": [
          "paperspaceUtils"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "send user list"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "Paperspace API Error"
          }
        }
      }
    },
    "/access/give-access-from-email": {
      "post": {
        "description": "gives access to machine",
        "tags": [
          "paperspaceAccess"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "description": "Request body for giving access",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/EnableAccessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successfully granted access"
          },
          "400": {
            "description": "Bad request body"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "Paperspace API Error"
          }
        }
      }
    },
    "/access/revoke-access-from-email": {
      "post": {
        "description": "revokes access to machine",
        "tags": [
          "paperspaceAccess"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "description": "Optional description in *Markdown*",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/EnableAccessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successfully granted access"
          },
          "400": {
            "description": "Bad request body"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "501": {
            "description": "Paperspace API Error"
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "description": "login to change access",
        "produces": [
          "application/json"
        ],
        "tags": [
          "access"
        ],
        "requestBody": {
          "description": "Body used to log in as admin",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successfully logged in"
          },
          "400": {
            "description": "Bad request body"
          },
          "401": {
            "description": "not authorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "logs in and gets token as guest with limited functionality",
        "tags": [
          "access"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "got token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "EnableAccessRequest": {
      "type": "object",
      "required": [
        "name",
        "email"
      ],
      "properties": {
        "name": {
          "type": "string",
          "format": "name",
          "description": "Your name",
          "example": "Your name"
        },
        "email": {
          "type": "string",
          "format": "email",
          "description": "Email address associated with paperspace account",
          "example": "yourpaperspace@email.com"
        }
      }
    },
    "LoginRequest": {
      "type": "object",
      "required": [
        "name",
        "password"
      ],
      "properties": {
        "name": {
          "type": "string",
          "format": "name",
          "description": "Your username",
          "example": "your username"
        },
        "password": {
          "type": "string",
          "format": "password",
          "description": "password",
          "example": "password"
        }
      }
    },
...
  },
...
}